added main name
