+++ src/index.css
@@ -569,3 +569,4 @@ button:hover {
 
 /*Invoice page css end*/
 
+/*Item rate page css*/
